+++ src/app/features/committee-account-settings/committee-account-settings.component.html
@@ -8,7 +8,7 @@
     </div>
     <h2 class="mt-6 leading-[30px] font-extrabold">Account Settings</h2>
     <p class="mt-[22px]">
-      Edit or change your account informaton and password.
+      Edit or change your account information and password.
     </p>
     <dp-basic-card class="mt-8">
       <div class="w-full flex flex-col gap-11">
@@ -37,7 +37,9 @@
         <div
           class="w-full flex flex-col md:flex-row justify-between items-center"
         >
-          <p class="font-bold leading-5 md:min-w-[150px]">Contact informaton</p>
+          <p class="font-bold leading-5 md:min-w-[150px]">
+            Contact information
+          </p>
           <p class="max-w-full">{{ name }}</p>
           <button
             mat-flat-button
