+++ src/app/guards/exports.ts
@@ -1,3 +1,3 @@
 export * from './committee.guard';
-export * from './auth.guard';
+export * from './noAuth.guard';
 export * from './city-staff.guard';
