+++ src/app/logic/auth/sign-up.service.ts
@@ -12,7 +12,7 @@ import {
   Subject,
   tap,
 } from 'rxjs';
-import { AccountService } from 'src/app/auth/account-service/account.service';
+import { AccountService } from 'src/app/core/account-service/account.service';
 import { Result, SignUpCredentials } from 'src/app/shared/models/exports';
 
 @Injectable()
